requirements.txt
python-telegram-bot==12.0.0 
retrying==1.3.3 
google-cloud-firestore>=1.6.0 
ujson==1.35 
requests==2.20.1 
jinja2==2.10.1
pytest==5.3.2
pytest-mock==1.13.0
markdown2>=2.3.9